Switched from float32 to float64

+++ igrf/igrf.go
@@ -92,20 +92,20 @@ func IGRF(lat, lon, alt, date float64) (IGRFresults, error) {
 	// }
 
 	res := IGRFresults{
-		Declination:         float32(d),
-		DeclinationSV:       float32(ddot),
-		Inclination:         float32(i),
-		InclinationSV:       float32(idot),
-		HorizontalIntensity: float32(h),
-		HorizontalSV:        float32(hdot),
-		NorthComponent:      float32(x),
-		NorthSV:             float32(xdot),
-		EastComponent:       float32(y),
-		EastSV:              float32(ydot),
-		VerticalComponent:   float32(z),
-		VerticalSV:          float32(zdot),
-		TotalIntensity:      float32(f),
-		TotalSV:             float32(fdot),
+		Declination:         d,
+		DeclinationSV:       ddot,
+		Inclination:         i,
+		InclinationSV:       idot,
+		HorizontalIntensity: h,
+		HorizontalSV:        hdot,
+		NorthComponent:      x,
+		NorthSV:             xdot,
+		EastComponent:       y,
+		EastSV:              ydot,
+		VerticalComponent:   z,
+		VerticalSV:          zdot,
+		TotalIntensity:      f,
+		TotalSV:             fdot,
 	}
 	return res, nil
 }
